requirements.txt
pyautogen>=0.2.35
chromadb>=0.4.22
groq>=0.4.0
google-generativeai>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
spacy>=3.7.0
textblob>=0.17.1
python-dotenv>=1.0.0
pydantic>=2.5.0
newsapi-python>=0.2.7
ddgs>=1.0.0
scholarly>=1.7.11
arxiv>=2.1.0
pandas>=2.1.0
numpy>=1.24.0
aiohttp>=3.9.0
lxml>=4.9.0
html5lib>=1.1
python-dateutil>=2.8.0
openai>=1.0.0
